+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Added `sql_server_to_parquet` Prefect flow.
 - Added `sap_to_parquet` Prefect flow.
 - Added `duckdb_to_sql_server`, `duckdb_to_parquet`, `duckdb_transform` Prefect flows.
 - Added `bcp` and `duckdb_query` Prefect tasks.
@@ -71,6 +72,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed `catch_extra_separators()` from `SAPRFCV2` class
 
 ### Fixed
+- Fixed `bcp` prefect task to run correct.
 - Fixed the typo in credentials in `SQLServer` source
 
 ## [0.4.3] - 2022-04-28